pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ramalama"
dynamic = ["version"]
description = "RamaLama is a command line tool for working with AI LLM models."
readme = "README.md"
requires-python = ">=3.10"
keywords = [
  "ramalama",
  "llama",
  "ai",
  "llm",
  "large-language-model",
  "inference",
  "containers",
  "podman",
  "docker",
  "gpu",
]
dependencies = [
  "argcomplete",
  "pyyaml",
  "jsonschema",
  "jinja2",
]
maintainers = [
   { name = "Dan Walsh", email = "dawalsh@redhat.com" },
   { name = "Mike Bonnet", email = "mikeb@redhat.com" },
]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Intended Audience :: System Administrators",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Utilities",
]

[project.license]
text = "MIT"

[project.optional-dependencies]
dev = [
    "argcomplete~=3.0",
    "codespell~=2.0",
    "ruff>=0.14.14",
    "mypy",
    "types-PyYAML",
    "types-jsonschema",
]
rag = [
    "docling-core",
    "Pillow",
    "pypdfium2",
    "qdrant-client[fastembed]",
]

[project.urls]
Homepage = "https://github.com/containers/ramalama"
Repository = "https://github.com/containers/ramalama"
Issues = "https://github.com/containers/ramalama/issues"
Discussions = "https://github.com/containers/ramalama/discussions"
Discord = "https://discord.gg/MkCXuTRBUn"
Matrix = "https://matrix.to/#/#ramalama:fedoraproject.org"

[project.scripts]
ramalama = "ramalama.cli:main"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.package-data]
"ramalama" = ["py.typed"]

[tool.setuptools.dynamic]
version = {attr = "ramalama.version.version"}

[tool.codespell]
skip = ["build", "ramalama.egg-info", "logos", ".git", "venv", ".venv", ".tox", ".pixi"]
dictionary = ".codespelldict"
ignore-words-list = [
  "assertin",
  "cann",
  "clos",
  "creat",
  "hastable",
  "passt",
  "ro",
  "shouldnot",
]
check-hidden = true

[tool.ruff]
line-length = 120
target-version = "py310"
include = ["\\.pyi?$"]
extend-include = ["bin/ramalama"]
exclude = [
  "/\\.git",
  "/\\.tox",
  "/\\.venv",
  "/\\.history",
  "/build",
  "/dist",
  "/docs",
  "/hack",
  "/venv",
  "/.pixi"
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.format]
preview = true
quote-style = "preserve"

[tool.setuptools.packages.find]
include = ["ramalama", "ramalama.*"]

[tool.setuptools.data-files]
"share/ramalama" = ["shortnames/shortnames.conf", "docs/ramalama.conf"]
"share/ramalama/inference" = [
  "inference-spec/schema/*.json",
  "inference-spec/engines/*",
]
"share/man/man1" = ["docs/*.1"]
"share/man/man5" = ["docs/*.5"]
"share/man/man7" = ["docs/*.7"]
"share/bash-completion/completions" = ["completions/bash-completion/completions/*"]
"share/zsh/site-functions" = ["completions/zsh/site-functions/*"]
"share/fish/vendor_completions.d" = ["completions/fish/vendor_completions.d/*"]

[tool.check-wheel-contents]
# W002 fails on duplicate files, and completions for bash and zsh are identical
ignore = ["W002"]

[tool.mypy]
python_version = "3.10"